+++ src/common/const/index.ts
@@ -1,0 +1,2 @@
+export { UnivMailMap } from './univMailMap.const';
+export { HobbyMask, CharacterMask } from './mask.const';
