+++ src/common/const/mask.const.ts
@@ -1,0 +1,9 @@
+export const HobbyMask = {
+  Reading: 0x1,
+};
+
+export const CharacterMask = {
+  Cute: 0x1,
+  Extroverted: 0x2,
+  Introverted: 0x4,
+};
