+++ src/common/enums/character.enum.ts
@@ -3,9 +3,3 @@ export enum Character {
   Extroverted = '외향적인',
   Introverted = '내향적인',
 }
-
-export const CharacterMask = {
-  Cute: 0x1,
-  Extroverted: 0x2,
-  Introverted: 0x4,
-};
