+++ src/common/enums/hobby.enum.ts
@@ -1,7 +1,3 @@
 export enum Hobby {
   Reading = '독서',
 }
-
-export const HobbyMask = {
-  Reading: 0x1,
-};
