+++ src/common/enums/index.ts
@@ -1,0 +1,10 @@
+export { Character } from './character.enum';
+export { Degree } from './degree.enum';
+export { Hobby } from './hobby.enum';
+export { Major } from './major.enum';
+export { Mbti } from './mbti.enum';
+export { Nickname } from './nickname.enum';
+export { Page } from './page.enum';
+export { Sex } from './sex.enum';
+export { SexOrient } from './sexOrient.enum';
+export { University } from './university.enum';
