+++ src/common/enums/mbti.enum.ts
@@ -1,4 +1,3 @@
-/* eslint-disable no-unused-vars */
 export enum Mbti {
   INFP = 'infp',
   INFJ = 'infj',
