+++ src/common/enums/sex.enum.ts
@@ -1,4 +1,3 @@
-/* eslint-disable no-unused-vars */
 export enum Sex {
   Female = 'F',
   Male = 'M',
