+++ src/common/enums/sexOrient.enum.ts
@@ -1,4 +1,3 @@
-/* eslint-disable no-unused-vars */
 export enum SexOrient {
   Heterosexual = 'hetero',
   Homosexual = 'homo',
