+++ src/common/enums/university.enum.ts
@@ -1,4 +1,3 @@
-/* eslint-disable no-unused-vars */
 export enum University {
   가천대학교 = '가천대학교',
   가톨릭상지대학교 = '가톨릭상지대학교',
@@ -193,419 +192,3 @@ export enum University {
   고려사이버대학교 = 'cuk.edu',
   */
 }
-
-export const UnivMap = {
-  가천대학교: 'gachon.ac.kr',
-  가톨릭상지대학교: 'csj.ac.kr',
-  강동대학교: 'gangdong.ac.kr',
-  강릉영동대학교: 'gyc.ac.kr',
-  강원관광대학: 'kt.ac.kr',
-  강원도립대학: 'gw.ac.kr',
-  거제대학교: 'koje.ac.kr',
-  경기과학기술대학교: 'gtec.ac.kr',
-  경남도립거창대학: 'gc.ac.kr',
-  경남도립남해대학: 'namhae.ac.kr',
-  경남정보대학교: 'kit.ac.kr',
-  경민대학교: 'kyungmin.ac.kr',
-  경복대학교: 'kbu.ac.kr',
-  경북과학대학교: 'kbsc.ac.kr',
-  경북도립대학교: 'gpc.ac.kr',
-  경북전문대학교: 'kbc.ac.kr',
-  경원전문대학: 'kwc.ac.kr',
-  경인여자대학교: 'kic.ac.kr',
-  계명문화대학교: 'kmcu.ac.kr',
-  계원예술대학교: 'kaywon.ac.kr',
-  고구려대학교: 'kgrc.ac.kr',
-  광양보건대학교: 'kwangyang.ac.kr',
-  광주보건대학교: 'ghu.ac.kr',
-  구미대학교: 'gumi.ac.kr',
-  구세군사관학교: 'saotc.ac.kr',
-  국제대학교: 'kookje.ac.kr',
-  군산간호대학교: 'kcn.ac.kr',
-  군장대학교: 'kunjang.ac.kr',
-  기독간호대학교: 'ccn.ac.kr',
-  김천과학대학교: 'kcs.ac.kr',
-  김천대학교: 'gimcheon.ac.kr',
-  김포대학교: 'kimpo.ac.kr',
-  김해대학교: 'gimhae.ac.kr',
-  농협대학교: 'nonghyup.ac.kr',
-  대경대학교: 'tk.ac.kr',
-  대구공업대학교: 'ttc.ac.kr',
-  대구과학대학교: 'tsu.ac.kr',
-  대구미래대학교: 'dfc.ac.kr',
-  대구보건대학교: 'dhc.ac.kr',
-  대덕대학교: 'ddu.ac.kr',
-  대동대학교: 'daedong.ac.kr',
-  대림대학교: 'daelim.ac.kr',
-  대원대학교: 'daewon.ac.kr',
-  대전보건대학교: 'hit.ac.kr',
-  동강대학교: 'dkc.ac.kr',
-  동남보건대학교: 'dongnam.ac.kr',
-  동명대학교: 'tu.ac.kr',
-  동부산대학교: 'dpc.ac.kr',
-  동서울대학교: 'dsc.ac.kr',
-  동아방송예술대학교: 'dima.ac.kr',
-  동아인재대학교: 'dongac.ac.kr',
-  동양미래대학교: 'dongyang.ac.kr',
-  동우대학교: 'duc.ac.kr',
-  동원과학기술대학교: 'dist.ac.kr',
-  동원대학교: 'tw.ac.kr',
-  동의과학대학교: 'dit.ac.kr',
-  동주대학교: 'dongju.ac.kr',
-  두원공과대학교: 'doowon.ac.kr',
-  마산대학교: 'masan.ac.kr',
-  명지전문대학교: 'mjc.ac.kr',
-  목포과학대학교: 'mokpo-c.ac.kr',
-  문경대학교: 'mkc.ac.kr',
-  배화여자대학교: 'baewha.ac.kr',
-  백석문화대학교: 'bscu.ac.kr',
-  백제예술대학교: 'paekche.ac.kr',
-  벽성대학교: 'bs.ac.kr',
-  부산경상대학교: 'bsks.ac.kr',
-  부산과학기술대학교: 'bist.ac.kr',
-  부산여자대학교: 'bwc.ac.kr',
-  부산예술대학교: 'busanarts.ac.kr',
-  부천대학교: 'bc.ac.kr',
-  삼육보건대학: 'shu.ac.kr',
-  삼육의명대학: 'syu.ac.kr',
-  상지영서대학교: 'sy.ac.kr',
-  서라벌대학교: 'sorabol.ac.kr',
-  서영대학교: 'seoyeong.ac.kr',
-  서울보건대학교: 'shjc.ac.kr',
-  서울여자간호대학교: 'snjc.ac.kr',
-  서울예술대학교: 'seoularts.ac.kr',
-  서일대학교: 'seoil.ac.kr',
-  서정대학교: 'seojeong.ac.kr',
-  서해대학교: 'sohae.ac.kr',
-  선린대학교: 'sunlin.ac.kr',
-  성덕대학교: 'sdc.ac.kr',
-  성심외국어대학교: 'sungsim.ac.kr',
-  세경대학교: 'saekyung.ac.kr',
-  송곡대학교: 'songgok.ac.kr',
-  송원대학교: 'songwon.ac.kr',
-  송호대학교: 'songho.ac.kr',
-  수성대학교: 'sc.ac.kr',
-  수원과학대학교: 'ssc.ac.kr',
-  수원여자대학교: 'swc.ac.kr',
-  순천제일대학교: 'suncheon.ac.kr',
-  숭의여자대학교: 'sewc.ac.kr',
-  신구대학교: 'shingu.ac.kr',
-  신성대학교: 'shinsung.ac.kr',
-  신안산대학교: 'sau.ac.kr',
-  신흥대학교: 'shc.ac.kr',
-  아주자동차대학: 'motor.ac.kr',
-  안동과학대학교: 'asc.ac.kr',
-  안산대학교: 'ansan.ac.kr',
-  여주대학교: 'yit.ac.kr',
-  연성대학교: 'yeonsung.ac.kr',
-  연암공과대학교: 'yc.ac.kr',
-  영남외국어대학: 'yflc.ac.kr',
-  영남이공대학교: 'ync.ac.kr',
-  영진사이버대학: 'ycc.ac.kr',
-  영진전문대학: 'yjc.ac.kr',
-  오산대학교: 'osan.ac.kr',
-  용인송담대학교: 'ysc.ac.kr',
-  우송공업대학: 'wst.ac.kr',
-  우송정보대학: 'wsi.ac.kr',
-  울산과학대학교: 'uc.ac.kr',
-  웅지세무대학: 'wat.ac.kr',
-  원광보건대학교: 'wkhc.ac.kr',
-  원주대학: 'wonju.ac.kr',
-  유한대학교: 'yuhan.ac.kr',
-  인덕대학교: 'induk.ac.kr',
-  인천재능대학교: 'jeiu.ac.kr',
-  인천전문대학: 'icc.ac.kr',
-  인하공업전문대학: 'itc.ac.kr',
-  장안대학교: 'jangan.ac.kr',
-  적십자간호대학: 'cau.ac.kr',
-  전남과학대학교: 'chunnam-c.ac.kr',
-  전남도립대학교: 'dorip.ac.kr',
-  전북과학대학교: 'jbsc.ac.kr',
-  전주기전대학: 'jk.ac.kr',
-  전주비전대학교: 'jvision.ac.kr',
-  제주관광대학교: 'ctc.ac.kr',
-  제주산업정보대학: 'jeju.ac.kr',
-  제주한라대학교: 'chu.ac.kr',
-  조선간호대학교: 'cnc.ac.kr',
-  조선이공대학교: 'cst.ac.kr',
-  진주보건대학교: 'jhc.ac.kr',
-  창신대학: 'csc.ac.kr',
-  창원문성대학: 'cmu.ac.kr',
-  천안연암대학: 'yonam.ac.kr',
-  청강문화산업대학교: 'chungkang.academy',
-  청암대학교: 'scjc.ac.kr',
-  춘해보건대학교: 'ch.ac.kr',
-  충남도립청양대학: 'cyc.ac.kr',
-  충북도립대학: 'cpu.ac.kr',
-  충북보건과학대학교: 'chsu.ac.kr',
-  충청대학교: 'ok.ac.kr',
-  포항대학교: 'pohang.ac.kr',
-  한국골프대학: 'kg.ac.kr',
-  한국관광대학교: 'ktc.ac.kr',
-  한국농수산대학: 'af.ac.kr',
-  한국복지대학교: 'hanrw.ac.kr',
-  한국복지사이버대학: 'corea.ac.kr',
-  한국승강기대학교: 'klc.ac.kr',
-  한국영상대학교: 'pro.ac.kr',
-  한국정보통신기능대학: 'icpc.ac.kr',
-  한국철도대학: 'krc.ac.kr',
-  한국폴리텍대학교: 'kopo.ac.kr',
-  한림성심대학교: 'hsc.ac.kr',
-  한양여자대학교: 'hywoman.ac.kr',
-  한영대학: 'hanyeong.ac.kr',
-  혜전대학: 'hj.ac.kr',
-  혜천대학교: 'hu.ac.kr',
-  가야대학교: 'kaya.ac.kr',
-  가천의과학대학교: 'gachon.ac.kr',
-  가톨릭대학교: 'catholic.ac.kr',
-  감리교신학대학교: 'mtu.ac.kr',
-  강남대학교: 'kangnam.ac.kr',
-  강릉원주대학교: 'gwnu.ac.kr',
-  강원대학교: 'kangwon.ac.kr',
-  건국대학교: 'konkuk.ac.kr',
-  건국대학교글로컬: 'kku.ac.kr',
-  건양대학교: 'konyang.ac.kr',
-  건양사이버대학교: 'kycu.ac.kr',
-  경기대학교: 'kyonggi.ac.kr',
-  경남과학기술대학교: 'gntech.ac.kr',
-  경남대학교: 'hanma.kr',
-  경동대학교: 'k1.ac.kr',
-  경북대학교: 'knu.ac.kr',
-  경북외국어대학교: 'kufs.ac.kr',
-  경상대학교: 'gnu.ac.kr',
-  경성대학교: 'ks.ac.kr',
-  경운대학교: 'ikw.ac.kr',
-  경운대학교산업대: 'ikw.ac.kr',
-  경인교육대학교: 'ginue.ac.kr',
-  경일대학교: 'kiu.kr',
-  경주대학교: 'gju.ac.kr',
-  경희대학교: 'khu.ac.kr',
-  경희사이버대학교: 'khcu.ac.kr',
-  계명대학교: 'kmu.ac.kr',
-  고려대학교: 'korea.ac.kr',
-  고려대학교세종: 'korea.ac.kr',
-  고려사이버대학교: 'cuk.edu',
-  /*
-고신대학교 kosin.ac.kr
-공주교육대학교 gjue.ac.kr
-공주대학교 smail.kongju.ac.kr
-가톨릭관동대학교 cku.ac.kr
-광신대학교 kwangshin.ac.kr
-광운대학교 kw.ac.kr
-광주가톨릭대학교 kjcatholic.ac.kr
-GIST gist.ac.kr
-광주교육대학교 gnue.ac.kr
-광주대학교 gwangju.ac.kr
-광주대학교(산업대) gwangju.ac.kr
-광주여자대학교 kwu.ac.kr
-국민대학교 kookmin.ac.kr
-국제사이버대학교 gcu.ac
-군산대학교 kunsan.ac.kr
-그리스도대학교 kcu.ac.kr
-극동대학교 kdu.ac.kr
-글로벌사이버대학교 global.ac.kr
-금강대학교 ggu.ac.kr
-금오공과대학교 kumoh.ac.kr
-김천대학교 gimcheon.ac.kr
-꽃동네대학교 kkot.ac.kr
-나사렛대학교 kornu.ac.kr
-남부대학교 nambu.ac.kr
-남서울대학교 nsu.ac.kr
-남서울대학교(산업대) nsu.ac.kr
-단국대학교 dankook.ac.kr
-대구가톨릭대학교 cu.ac.kr
-DGIST dgist.ac.kr
-대구교육대학교 dnue.ac.kr
-대구대학교 daegu.ac.kr
-대구사이버대학교 dcu.ac.kr
-대구예술대학교 dgau.ac.kr
-대구외국어대학교 dufs.ac.kr
-대구한의대학교 dhu.ac.kr
-대신대학교 daeshin.ac.kr
-대전가톨릭대학교 dcatholic.ac.kr
-대전대학교 edu.dju.ac.kr
-대전신학교 daejeon.ac.kr
-대전신학대학교 daejeon.ac.kr
-대진대학교 daejin.ac.kr
-덕성여자대학교 duksung.ac.kr
-동국대학교 dongguk.edu
-동국대학교(경주) dongguk.ac.kr
-동덕여자대학교 dongduk.ac.kr
-동명대학교 tu.ac.kr
-동명정보대학교 tu.ac.kr
-동서대학교 dongseo.ac.kr
-동신대학교 dsu.kr
-동아대학교 donga.ac.kr
-동양대학교 dyu.ac.kr
-동의대학교 deu.ac.kr
-디지털서울문화예술대학교 scau.ac.kr
-루터대학교 ltu.ac.kr
-명지대학교 mju.ac.kr
-목원대학교 mokwon.ac.kr
-목포가톨릭대학교 mcu.ac.kr
-목포대학교 mokpo.ac.kr
-목포해양대학교 mmu.ac.kr
-배재대학교 pcu.ac.kr
-백석대학교 bu.ac.kr
-부경대학교 pukyong.ac.kr
-부산가톨릭대학교 cup.ac.kr
-부산교육대학교 bnue.ac.kr
-부산대학교 pusan.ac.kr
-부산디지털대학교 bdu.ac.kr
-부산외국어대학교 bufs.ac.kr
-부산장신대학교 bpu.ac.kr
-사이버한국외국어대학교 cufs.ac.kr
-삼육대학교 syuin.ac.kr
-상명대학교 sangmyung.kr
-상명대학교(천안) sangmyung.kr
-상주대학교 knu.ac.kr
-상지대학교 sangji.ac.kr
-서강대학교 sogang.ac.kr
-서경대학교 skuniv.ac.kr
-서남대학교 seonam.ac.kr
-서울과학기술대학교 seoultech.ac.kr
-서울과학기술대학교(산업대) seoultech.ac.kr
-서울교육대학교 snue.ac.kr
-서울기독대학교 scu.ac.kr
-서울대학교 snu.ac.kr
-서울디지털대학교 sdu.ac.kr
-서울사이버대학교 iscu.ac.kr
-서울시립대학교 uos.ac.kr
-서울신학대학교 stu.ac.kr
-서울여자대학교 swu.ac.kr
-서울장신대학교 sjs.ac.kr
-서원대학교 seowon.ac.kr
-선문대학교 sunmoon.ac.kr
-성결대학교 sungkyul.ac.kr
-성공회대학교 skhu.ac.kr
-성균관대학교 skku.edu
-성신여자대학교 sungshin.ac.kr
-세명대학교 semyung.ac.kr
-세종대학교 sju.ac.kr
-세종사이버대학교 sjcu.ac.kr
-세한대학교 sehan.ac.kr
-송원대학교 songwon.ac.kr
-수원가톨릭대학교 suwoncatholic.ac.kr
-수원대학교 suwon.ac.kr
-숙명여자대학교 sookmyung.ac.kr
-순복음총회신학교 kcc.ac.kr
-순천대학교 scnu.ac.kr
-순천향대학교 sch.ac.kr
-숭실대학교 soongsil.ac.kr
-숭실사이버대학교 kcu.ac
-신경대학교 sgu.ac.kr
-신라대학교 silla.ac.kr
-아세아연합신학대학교 acts.ac.kr
-아주대학교 ajou.ac.kr
-안동대학교 anu.ac.kr
-안양대학교 ayum.anyang.ac.kr
-연세대학교 yonsei.ac.kr
-연세대학교(원주) yonsei.ac.kr
-열린사이버대학교 ocu.ac.kr
-영남대학교 ynu.ac.kr
-영남신학대학교 ytus.ac.kr
-유원대학교 u1.ac.kr
-영산대학교 ysu.ac.kr
-영산대학교(산업대) ysu.ac.kr
-영산선학대학교 youngsan.ac.kr
-예수대학교 jesus.ac.kr
-예원예술대학교 yewon.ac.kr
-용인대학교 yiu.ac.kr
-우석대학교 woosuk.ac.kr
-우송대학교 wsu.ac.kr
-우송대학교(산업대) wsu.ac.kr
-UNIST unist.ac.kr
-울산대학교 ulsan.ac.kr
-원광대학교 wonkwang.ac.kr
-원광디지털대학교 wdu.ac.kr
-위덕대학교 uu.ac.kr
-을지대학교 eulji.ac.kr
-이화여자대학교 ewhain.net
-인제대학교 inje.ac.kr
-인천가톨릭대학교 iccu.ac.kr
-인천대학교 inu.ac.kr
-인하대학교 inha.edu
-장로회신학대학교 pcts.ac.kr
-전남대학교 jnu.ac.kr
-전북대학교 jbnu.ac.kr
-전주교육대학교 jnue.kr
-전주대학교 jj.ac.kr
-정석대학 jit.ac.kr
-제주교육대학교 jejue.ac.kr
-제주국제대학교 jeju.ac.kr
-제주대학교 jejunu.ac.kr
-조선대학교 chosun.kr
-중부대학교 jmail.ac.kr
-중앙대학교 cau.ac.kr
-중앙대학교(안성) cau.ac.kr
-중앙승가대학교 sangha.ac.kr
-중원대학교 jwu.ac.kr
-진주교육대학교 cue.ac.kr
-진주산업대학교(산업대) gntech.ac.kr
-차의과학대학교 cha.ac.kr
-창신대학교 cs.ac.kr
-창원대학교 changwon.ac.kr
-청운대학교 chungwoon.ac.kr
-청주교육대학교 cje.ac.kr
-청주대학교 cju.ac.kr
-초당대학교 chodang.ac.kr
-초당대학교(산업대) chodang.ac.kr
-총신대학교 chongshin.ac.kr
-추계예술대학교 chugye.ac.kr
-춘천교육대학교 cnue.ac.kr
-충남대학교 cnu.ac.kr
-충북대학교 chungbuk.ac.kr
-침례신학대학교 kbtus.ac.kr
-칼빈대학교 calvin.ac.kr
-탐라대학교 tnu.ac.kr
-평택대학교 ptu.ac.kr
-POSTECH postech.ac.kr
-한경대학교 hknu.ac.kr
-한경대학교(산업대) hknu.ac.kr
-KAIST kaist.ac.kr
-한국교원대학교 knue.ac.kr
-한국교통대학교 ut.ac.kr
-한국교통대학교(산업대) ut.ac.kr
-한국국제대학교 iuk.ac.kr
-한국기술교육대학교 koreatech.ac.kr
-한국방송통신대학교 knou.ac.kr
-한국산업기술대학교 kpu.ac.kr
-한국산업기술대학교(산업대) kpu.ac.kr
-한국성서대학교 bible.ac.kr
-한국예술종합학교 karts.ac.kr
-한국외국어대학교 hufs.ac.kr
-한국전통문화대학교 nuch.ac.kr
-한국체육대학교 knsu.ac.kr
-한국항공대학교 kau.kr
-한국해양대학교 kmou.ac.kr
-한남대학교 hannam.ac.kr
-한동대학교 handong.edu
-한라대학교 halla.ac.kr
-한려대학교 hanlyo.ac.kr
-한려대학교(산업대) hanlyo.ac.kr
-한림대학교 hallym.ac.kr
-한민학교 hanmin.ac.kr
-한밭대학교 hanbat.ac.kr
-한밭대학교(산업대) hanbat.ac.kr
-한북대학교 hanbuk.ac.kr
-한서대학교 hanseo.ac.kr
-한성대학교 hansung.ac.kr
-한세대학교 uohs.ac.kr
-한신대학교 hs.ac.kr
-한양대학교 hanyang.ac.kr
-한양대학교(ERICA) hanyang.ac.kr
-한양사이버대학교 hycu.ac.kr
-한영신학대학교 hytu.ac.kr
-한일장신대학교 hanil.ac.kr
-한중대학교 hanzhong.ac.kr
-협성대학교 uhs.ac.kr
-호남대학교 honam.ac.kr
-호남신학대학교 htus.ac.kr
-호산대학교 hosan.ac.kr
-호서대학교 hoseo.edu
-호원대학교 howon.ac.kr
-홍익대학교 hongik.ac.kr
-홍익대학교(세종) hongik.ac.kr
-화신사이버대학교: 'hscu.ac.kr'
-*/
-};
