+++ src/entities/index.ts
@@ -1,0 +1,5 @@
+export { AuthMailEntity } from './authMail.entity';
+export { AuthPhoneNumberEntity } from './authPhoneNumber.entity';
+export { UserImageEntity } from './userImage.entity';
+export { UserInfoEntity } from './userInfo.entity';
+export { UserEntity } from './users.entity';
