+++ src/user/user.module.ts
@@ -1,9 +1,8 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { UserEntity } from 'src/entities/users.entity';
+import { UserEntity, UserInfoEntity } from 'src/entities';
 import { UserController } from './user.controller';
 import { UserService } from './user.service';
-import { UserInfoEntity } from 'src/entities/userInfo.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([UserEntity, UserInfoEntity])],
